update: validate 调整
